Fix config file collection.
Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: pr test_setup_vm DdbTest DaosCoreTestDfuse

Required-githooks: true
